:recycle: refactor: alterar estrutura de pedidos

+++ src/main/java/com/mobile/pontoGestao/Services/PedidosService.java
@@ -2,21 +2,27 @@ package com.mobile.pontoGestao.Services;
 
 import com.google.api.core.ApiFuture;
 import com.google.cloud.firestore.Firestore;
+import com.google.cloud.firestore.Query;
 import com.google.cloud.firestore.QuerySnapshot;
 import com.mobile.pontoGestao.Dtos.Request.PedidoRequest;
 import com.mobile.pontoGestao.Dtos.Request.PedidoRequestUpdate;
+import com.mobile.pontoGestao.Dtos.Request.SenhaRequest;
 import com.mobile.pontoGestao.Dtos.Response.PedidoResponse;
 import com.mobile.pontoGestao.Enums.OrdenacaoPedido;
 import com.mobile.pontoGestao.Enums.StatusPedido;
 import com.mobile.pontoGestao.Enums.TipoPedido;
 import com.mobile.pontoGestao.Erros.EntityNotFoundException;
+import com.mobile.pontoGestao.Erros.LoginInvalidException;
 import com.mobile.pontoGestao.Mappers.ItemsPedidoMapper;
 import com.mobile.pontoGestao.Mappers.PedidosMapper;
 import com.mobile.pontoGestao.Models.Clientes;
 import com.mobile.pontoGestao.Models.ItemsPedido;
 import com.mobile.pontoGestao.Models.Pedidos;
+import com.mobile.pontoGestao.Models.Usuarios;
 import lombok.RequiredArgsConstructor;
 import com.google.cloud.Timestamp;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -32,6 +38,7 @@ public class PedidosService {
     private final ItemsPedidoMapper itemsPedidoMapper;
     private final PedidosMapper pedidosMapper;
     private final Firestore firestore;
+    private final PasswordEncoder passwordEncoder;
 
     public PedidoResponse criarPedido(PedidoRequest request) throws ExecutionException, InterruptedException {
         List<ItemsPedido> itemsPedidos = request.itens().stream().map(itemsPedidoMapper::toModel).toList();
@@ -49,15 +56,6 @@ public class PedidosService {
         pedido.setOrcamento(orcamento);
         pedido.setNomeCliente(cliente.getNome());
 
-        if (request.dataProva() != null) {
-            pedido.setDataProva(convertTimestamp(request.dataProva()));
-        }
-
-        if (request.dataEntrega() != null) {
-            pedido.setDataEntrega(convertTimestamp(request.dataEntrega()));
-        }
-
-        pedido.setDataPrazo(convertTimestamp(request.dataPrazo()));
         firestore.collection("pedidos").document(pedido.getId()).set(pedido);
         return pedidosMapper.toDto(pedido);
     }
@@ -69,17 +67,21 @@ public class PedidosService {
         return pedidosMapper.toDto(pedido);
     }
 
-    public List<PedidoResponse> verPedidos(StatusPedido statusPedido, TipoPedido tipoPedido, String titulo, OrdenacaoPedido ordenacao) throws ExecutionException, InterruptedException {
-        ApiFuture<QuerySnapshot> future =  firestore.collection("pedidos")
-                .get();
+    public List<PedidoResponse> verPedidos(StatusPedido statusPedido, String titulo, OrdenacaoPedido ordenacao, LocalDateTime inicioPrazo, LocalDateTime fimPrazo) throws ExecutionException, InterruptedException {
 
-        QuerySnapshot snapshot = future.get();
+
+        Query query = firestore.collection("pedidos");
+
+        if (statusPedido != null) {
+            query = query.whereEqualTo("statusPedido", statusPedido);
+        }
+
+        QuerySnapshot snapshot = query.get().get();
+
         Comparator<PedidoResponse> comparator = Comparator.comparing(PedidoResponse::dataPrazo);
-
-        if (ordenacao != null)
-        {
+        if (ordenacao != null) {
             comparator = switch (ordenacao) {
-                case  NOME-> Comparator.comparing(PedidoResponse::nomeCliente);
+                case NOME -> Comparator.comparing(PedidoResponse::nomeCliente);
                 case TITULO -> Comparator.comparing(PedidoResponse::titulo);
                 default -> Comparator.comparing(PedidoResponse::dataPrazo);
             };
@@ -89,8 +91,6 @@ public class PedidosService {
                 .stream()
                 .map(doc -> doc.toObject(Pedidos.class))
                 .map(pedidosMapper::toDto)
-                .filter(pedido -> statusPedido == null || statusPedido.equals(pedido.statusPedido()))
-                .filter(pedido -> tipoPedido == null || tipoPedido.equals(pedido.tipoPedido()))
                 .filter(pedido -> titulo == null || pedido.titulo().toLowerCase().contains(titulo.toLowerCase()))
                 .sorted(comparator)
                 .toList();
@@ -121,21 +121,18 @@ public class PedidosService {
             }
         }
 
-        if (request.dataProva() != null) {
-            pedido.setDataProva(convertTimestamp(request.dataProva()));
-        }
-
-        if (request.dataEntrega() != null) {
-            pedido.setDataEntrega(convertTimestamp(request.dataEntrega()));
-        }
-
-        if (request.dataPrazo() != null) {
-            pedido.setDataPrazo(convertTimestamp(request.dataPrazo()));
-        }
-
         pedidosMapper.updatePedido(request, pedido);
         firestore.collection("pedidos").document(pedido.getId()).set(pedido);
         return pedidosMapper.toDto(pedido);
+    }
+
+    public List<PedidoResponse> getPedidosByCliente(String idCliente) throws ExecutionException, InterruptedException {
+        QuerySnapshot snapshot = firestore.collection("pedidos").whereEqualTo("idCliente", idCliente).get().get();
+        return snapshot.getDocuments()
+                .stream()
+                .map(doc -> doc.toObject(Pedidos.class))
+                .map(pedidosMapper::toDto)
+                .toList();
     }
 
     private QuerySnapshot getDocumentSnapshots(String colecao, String id, String message) throws InterruptedException, ExecutionException {
@@ -148,7 +145,19 @@ public class PedidosService {
         return snapshot;
     }
 
-    public void deletarPedidos(String id) {
+    public void deletarPedidos(String id, SenhaRequest request) throws ExecutionException, InterruptedException {
+        String idUsuario = SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getPrincipal().toString();
+        ApiFuture<QuerySnapshot> future = firestore.collection("usuarios")
+                .whereEqualTo("id", idUsuario)
+                .get();
+        QuerySnapshot snapshot = future.get();
+        if (snapshot.isEmpty()) throw new EntityNotFoundException("Não foi possivel encontrar o usuário");
+        Usuarios usuario = snapshot.getDocuments().getFirst().toObject(Usuarios.class);
+
+        if (!passwordEncoder.matches(request.senha(), usuario.getSenha()))  throw new LoginInvalidException("Não foi possivel validar a requisição");
         firestore.collection("pedidos").document(id).delete();
     }
 
@@ -159,6 +168,14 @@ public class PedidosService {
                                 .atZone(ZoneId.systemDefault())
                                 .toInstant()
                 )
+        );
+    }
+
+    private  Date convertDate(LocalDateTime horario) {
+        return Date.from(
+                horario
+                        .atZone(ZoneId.systemDefault())
+                        .toInstant()
         );
     }
 }
